Implemented SERIAL vs PARALLEL execution. Improved argument mapping. Implemented valueless parameters. Improved handling of building and executing commands.
You can now define a new variable on your ShellCommand classes: EXECUTION_MODE. You can set this to either:
 ShellCommand::PARALLEL_EXECUTION or ShellCommand::SERIAL_EXECUTION to run in PARALLEL or SERIAL mode respectively.

Note that this assumes the way in which the command will be run:
- PARALLEL will run ONE command for all given targets.
- SERIAL will run ONE command for EACH target.

Examples,
traceroute only accepts one target per command, so you would define the variable as SERIAL in this class.
fping on the other hand accepts any number of targets, you would define the variable as PARALLEL in this class.

Failed validations will now throw an exception.