Remove parent from domain group.
This causes significant performance impact because it will load the full embedded relationship of a parent, which again includes its subdomains, which will again load its parents, ... you get the picture.

Looked for a way to just force it to return the URI /api/domains/:id, but that does not seem to be an option anymore (used to be, but was removed in https://github.com/api-platform/core/pull/1696).

This feature can be enabled through Symfony starting from version 4.1, which we are not currently at.

As such, the best solutions we currently have are:
- writing our own serializer (complex and we'd mostly be reinventing the wheel)
- omitting the $parent property from the group, and instead exposing it through /api/domains/:id/parent.

Signed-off-by: Kevin Reniers <Kevin.Reniers@cegeka.com>