Use latest doctrine/doctrine-bundle recipe
This requires master installs to explicitly add the serverVersion information to their DATABASE_URL. The install docs already mention that this is required.

I can pre-emptively update this on the fireping master to mariadb-10.1.26.

However, running database installs shouldn't be affected anyway as doctrine will try to guess the server version anyway.