Pin doctrine/inflector to <1.4
This is necessary to remove hundreds of deprecation notices being spammed into our log files continuously. This is a recent deprecation that won't be fixed until doctrine/orm:2.8.

After we've upgraded to doctrine/orm:2.8, we can safely remove this as a root requirement again.

I've already created an issue for this to remind us when the time comes.

Loading the index page now shows about 200 fewer deprecations.