Send an HTTP 409 Response for DirtyInputExceptions.
These occur for requests where the input data was dirty and could not be processed. These are irrecoverable since the slave will not be able to correct this. As a result of the current implementation, the probe dispatcher will continue to retry these transactions (because any request that triggers a non-200/non-409 response is retried forever.)