Ensure broken results are discarded by the ProbeDispatcher. (#284)
* Add DirtyInputException.php

* Throw DirtyInputException instead of \Exception when input data is dirty.

Examples of such cases:

- Expected sample count and actual sample count mismatch
- Essential data missing from request

* Send an HTTP 409 Response for DirtyInputExceptions.

These occur for requests where the input data was dirty and could not be processed. These are irrecoverable since the slave will not be able to correct this. As a result of the current implementation, the probe dispatcher will continue to retry these transactions (because any request that triggers a non-200/non-409 response is retried forever.)

* Change test to assert for HTTP 409 instead of HTTP 500.

Co-authored-by: Kevin R <22906111+cgkkevinr@users.noreply.github.com>
Co-authored-by: Jimmy Cleuren <jimmy.cleuren@gmail.com>