Mount a volume into ./var/rrd/test to test RrdCachedStorage.
This works locally, question is if it'll work on Travis. Basically, RrdCachedStorage *expects* the daemon to be running locally. Obviously, we aren't doing that on Travis nor on our local docker-compose stack, so this is a workaround. We first create a volume, then mount that named volume into both the php and rrdcached containers. As we create files into the rrdcached container, they'll be shared to the PHP container as-if they were local.

It took some blood, sweat and tears to get this working, and it's dirty as hell, but it works. We'll do something better at a later date. At least we don't have to remove the test this way just to stop it being flaky.