Fix test flakiness in RrdCachedStorage by mounting the rrdcached filesystem locally. (#300)
* Mount a volume into ./var/rrd/test to test RrdCachedStorage.

This works locally, question is if it'll work on Travis. Basically, RrdCachedStorage *expects* the daemon to be running locally. Obviously, we aren't doing that on Travis nor on our local docker-compose stack, so this is a workaround. We first create a volume, then mount that named volume into both the php and rrdcached containers. As we create files into the rrdcached container, they'll be shared to the PHP container as-if they were local.

It took some blood, sweat and tears to get this working, and it's dirty as hell, but it works. We'll do something better at a later date. At least we don't have to remove the test this way just to stop it being flaky.

* Remove the quotation marks.

Looking at other examples online, this may not be necessary.

* Add echo on the TRAVIS_BUILD_DIR

* Try /tmp/fireping/rrd/test instead of cache map

* Change /var/rrd/test to /tmp/fireping/rrd.

This should avoid the issue of ruining the cache directory.

* Adjust the volume path to the correct one.

* Add debugs

* Add some debugging.

* Add some debugging.

* Add some debugging.

* Try to run rrdcached as root

* Try to run rrdcached as root

* Try to run rrdcached as root

* Try to run rrdcached as root

* Add some documentation to .travis.yml and remove dump statements.

Co-authored-by: Kevin R <22906111+cgkkevinr@users.noreply.github.com>
Co-authored-by: Jimmy Cleuren <jimmy.cleuren@gmail.com>