Remove ansible/ directory.
This is unmaintained, not up-to-date, and all it does is start a docker container.

We might wish to reintroduce this at some point, but I don't want anyone to start depending on this before v1.0.

Besides, if we were ever to actively want to support this, we should extract it to its own repository apart from the main code repository.