Add a unique constraint to names within parents.
This adds two things:

- a database constraint preventing duplicate name entries within the same parent domain
- an application-level validation constraint that will generate error messages for attempts to do so.

Note that because for MySQL, two NULL values are -not- equal, this does not prevent multiple root domains with the same name on a database level (i.e. because of a race condition). We could avoid this by setting the parent_id column to NOT NULL and instead assuming the value 0 or -1 is "root domain".

That way, a created index would be 0-name or -1-name, which can constrained for uniqueness.

This closes #32 and #37.