Add versions to slaves (#354)
* Add version column to Slave

* Add Version domain object

* Add VersionReaderInterface and GitVersionReader

* Set the version from the stats endpoint

* Add code to pass version data to StatsManager.php

* Pass version to StatsManager

* Add VersionTest.php

* Add GitVersionReaderTest.php

* Add a ProcessFactoryInterface.php and decouple from Symfony Processes

This allows us to test the individual components without relying on *actual* Processes or *actual* Git repositories.

* Add GitVersionReader.php and its test.

* Use SymfonyProcessFactory for GitVersionReader in ProbeDispatcherCommand.php

* Add tests for Process code

* Add versions to slave index/detail pages.

* Add version to SlaveCrudController index/detail pages

* Ensure getVersion is a VersionInterface

* Remove DEFAULT '' from the migration

Doctrine doesn't seem to recognize it very well and we already default it on the Entity anyway.

* Remove VersionInterface.php

* Remove some other unnecessary functionality

* Implement __toString() to fix admin index/detail views

* Remove unused import

* Add CacheVersionReader.php

CacheVersionReader.php is used to read the running version once and then save it in the database. This will cause an initial (extended) load time the first time this happens. Should this become an issue, we can write a CacheWarmer that loads this on cache:warmup/cache:clear.

Co-authored-by: Kevin R <22906111+cgkkevinr@users.noreply.github.com>
Co-authored-by: Jimmy Cleuren <jimmy.cleuren@gmail.com>