refactor(slave/dispatcher): extract fetch configuration timer
I'm making a conscious choice of disregarding the executionOffset here. We originally added this because the master node was getting overwhelmed with requests. However, this already isn't relevant anymore as all slaves aren't being started at exactly the same time.

As such, I think this is a safe change to make.