Remove snc_redis.doctrine configuration key
This is not the recommended way of configuring this bundle, it should be done directly through doctrine configuration (where we already do this anyway).

A recent change in 2.4 started breaking invalid configurations like this, they likely didn't work already anyway.

See /snc/SncRedisBundle/issues/555#issuecomment-578283987 where it is already mentioned we shouldn't be doing this and that the configuration option will be removed in the future anyway.