refactor(dispatcher): extract FetchConfiguration timer
By doing it like this we ensure that this task will run every 120s from the moment the loop was started. This is great because we don't have to manage the `$now % 120 === $this->randomFactor` anymore. This was originally necessary because *all* slaves would start at the *exact same time*. Doing it like this means that all slaves are naturally offset by a few seconds anyway, since they won't all start at exactly the same time.

This explanation can be assumed for the following few commits as well.