fix(dispatcher): throw errors when we can't encode instructions as JSON
This is a "breaking" change from how it currently works, because it'll currently alert and try to continue. However, this is the *wrong* approach, since it causes issues to be hidden. This can only occur because of developer mistakes, so we *should* fail hard here and let it break and loudly announce that the code is broken and should be fixed.