refactor(composer): remove legacy extra.symfony-* parameters and extra.branch-alias
As a side effect, this also applied some changes that were introduced relatively recently to symfony/flex: it'll unpack all of the package packs that were previously added. That's why some of these dependencies were changed.

Nothing should've actually changed about the dependencies, just how it looks in this file.